Remove redundant code and refactor
