+++ packages/codegen/src/indexer.ts
@@ -10,9 +10,10 @@ import { Writable } from 'stream';
 import _ from 'lodash';
 
 import { getTsForSol } from './utils/type-mappings';
-import { Param, getBaseType, isArrayType } from './utils/types';
+import { Param } from './utils/types';
 import { MODE_ETH_CALL, MODE_STORAGE } from './utils/constants';
 import { getFieldType } from './utils/subgraph';
+import { getBaseType, isArrayType } from './utils/helpers';
 
 const TEMPLATE_FILE = './templates/indexer-template.handlebars';
 
@@ -37,7 +38,7 @@ export class Indexer {
    * @param returnType Return type for the query.
    * @param stateVariableType Type of the state variable in case of state variable query.
    */
-  addQuery (contract: string, mode: string, name: string, params: Array<Param>, typeName: string, stateVariableType?: string): void {
+  addQuery (contract: string, mode: string, name: string, params: Array<Param>, typeName: any, stateVariableType?: string): void {
     // Check if the query is already added.
     if (this._queries.some(query => query.name === name)) {
       return;
