+++ packages/codegen/src/utils/types.ts
@@ -10,16 +10,3 @@ export interface Param {
 export const reservedNames = new Set([
   'Symbol'
 ]);
-
-export const isElementaryType = (typeName: any): boolean => (typeName.type === 'ElementaryTypeName');
-export const isArrayType = (typeName: any): boolean => (typeName.type === 'ArrayTypeName');
-
-export const getBaseType = (typeName: any): string | undefined => {
-  if (isElementaryType(typeName)) {
-    return typeName.name;
-  } else if (isArrayType(typeName)) {
-    return getBaseType(typeName.baseTypeName);
-  } else {
-    return undefined;
-  }
-};
