+++ packages/codegen/src/visitor.ts
@@ -65,21 +65,11 @@ export class Visitor {
       } else {
         const typeName = node.returnParameters[0].typeName;
         switch (typeName.type) {
-          case 'ElementaryTypeName': {
-            this._schema.addQuery(name, params, typeName);
-            this._resolvers.addQuery(name, params, typeName);
+          case 'ElementaryTypeName':
             this._entity.addQuery(name, params, typeName);
             this._database.addQuery(name, params, typeName);
             this._client.addQuery(name, params, typeName);
-
-            assert(this._contract);
-            this._indexer.addQuery(this._contract.name, MODE_ETH_CALL, name, params, typeName);
-            break;
-          }
-
-          case 'UserDefinedTypeName':
-            errorMessage = `No support in codegen for user defined return type from method "${node.name}"`;
-            break;
+            // falls through
 
           case 'ArrayTypeName':
             this._schema.addQuery(name, params, typeName);
@@ -87,6 +77,10 @@ export class Visitor {
 
             assert(this._contract);
             this._indexer.addQuery(this._contract.name, MODE_ETH_CALL, name, params, typeName);
+            break;
+
+          case 'UserDefinedTypeName':
+            errorMessage = `No support in codegen for user defined return type from method "${node.name}"`;
             break;
 
           default:
